+++ CHANGELOG.md
@@ -117,6 +117,8 @@ After commit 543db92
 - Add node version check requiring node 16 or greater, else exit with error.
 - Update filenames in restart.sh. The restart.sh script will only work using credentials.json as configuration source.
 - Created alternate restart2.sh script for use in a server configured with environment variables.
+- Add configuration validation that cookie secret has been defined and user account has been configured.
+- Updated README.md for dual configuration.
 
 ## [v0.2.43](https://github.com/cotarr/irc-hybrid-client/releases/tag/v0.2.43) 2023-05-28